refactor : ui directory 구조 변경
