+++ app/src/main/java/woowacourse/shopping/ui/cart/cartAdapter/CartItemType.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.cart
+package woowacourse.shopping.ui.cart.cartAdapter
 
 import woowacourse.shopping.model.CartUIModel
 import woowacourse.shopping.model.ProductUIModel
