+++ app/src/main/java/woowacourse/shopping/ui/productdetail/ProductDetailContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.productdetail.contract
+package woowacourse.shopping.ui.productdetail
 
 import woowacourse.shopping.model.ProductUIModel
