+++ app/src/main/java/woowacourse/shopping/ui/shopping/productAdapter/ProductsItemType.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.shopping
+package woowacourse.shopping.ui.shopping.productAdapter
 
 import woowacourse.shopping.model.ProductUIModel
