+++ app/src/main/java/woowacourse/shopping/ui/shopping/productAdapter/viewHolder/ItemViewHolder.kt
@@ -1,8 +1,8 @@
-package woowacourse.shopping.ui.shopping.viewHolder
+package woowacourse.shopping.ui.shopping.productAdapter.viewHolder
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
-import woowacourse.shopping.ui.shopping.ProductsItemType
+import woowacourse.shopping.ui.shopping.productAdapter.ProductsItemType
 
 sealed class ItemViewHolder(view: View) : RecyclerView.ViewHolder(view) {
     abstract fun bind(productItemType: ProductsItemType)
